Replaced FLEncoder>>encodePositiveInteger: with #encodeUint32:
Made implementation of integer encoding and decoding methods compatible with larger word sizes, i.e., passing an arbritrary integer to one of the encoding methods will encode
exactly the defined range. The #encodeUintXXX methods now also correctly mask the highest byte to ensure that when the passed integer is negative and the system uses a larger
word size then the number of bytes that are being encoded, only the high byte is encoded (e.g. 256 is not a valid byte but a valid integer, representing -128 in two's complement).